[Generic] Прогресс бары и NanoUI для радио, вендоматов и факса

+++ code/game/turfs/simulated/walls_reinforced.dm
@@ -100,7 +100,7 @@
 		if(WT.remove_fuel(0,user))
 			user << "<span class='notice'>You start repairing the damage to [src].</span>"
 			playsound(src, 'sound/items/Welder.ogg', 100, 1)
-			if(do_after(user, max(5, damage / 5)) && WT && WT.isOn())
+			if(do_after(user, max(5, damage / 5), target = src) && WT && WT.isOn())
 				user << "<span class='notice'>You finish repairing the damage to [src].</span>"
 				take_damage(-damage)
 			return
@@ -126,13 +126,13 @@
 				user << "<span class='notice'>You begin removing the support lines.</span>"
 				playsound(src, 'sound/items/Screwdriver.ogg', 100, 1)
 
-				sleep(40)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,40,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( d_state == 1 && user.loc == T && user.get_active_hand() == W )
-					src.d_state = 2
-					src.icon_state = "r_wall-2"
-					user << "<span class='notice'>You remove the support lines.</span>"
+					if( d_state == 1 && user.loc == T && user.get_active_hand() == W )
+						src.d_state = 2
+						src.icon_state = "r_wall-2"
+						user << "<span class='notice'>You remove the support lines.</span>"
 				return
 
 			//REPAIRING (replacing the outer grille for cosmetic damage)
@@ -156,13 +156,13 @@
 					user << "<span class='notice'>You begin slicing through the metal cover.</span>"
 					playsound(src, 'sound/items/Welder.ogg', 100, 1)
 
-					sleep(60)
-					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !WT || !WT.isOn() || !T )	return
+					if(do_after(user,60,target = src))
+						if( !istype(src, /turf/simulated/wall/r_wall) || !user || !WT || !WT.isOn() || !T )	return
 
-					if( d_state == 2 && user.loc == T && user.get_active_hand() == WT )
-						src.d_state = 3
-						src.icon_state = "r_wall-3"
-						user << "<span class='notice'>You press firmly on the cover, dislodging it.</span>"
+						if( d_state == 2 && user.loc == T && user.get_active_hand() == WT )
+							src.d_state = 3
+							src.icon_state = "r_wall-3"
+							user << "<span class='notice'>You press firmly on the cover, dislodging it.</span>"
 				else
 					user << "<span class='notice'>You need more welding fuel to complete this task.</span>"
 				return
@@ -172,13 +172,13 @@
 				user << "<span class='notice'>You begin slicing through the metal cover.</span>"
 				playsound(src, 'sound/items/Welder.ogg', 100, 1)
 
-				sleep(40)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,60,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( d_state == 2 && user.loc == T && user.get_active_hand() == W )
-					src.d_state = 3
-					src.icon_state = "r_wall-3"
-					user << "<span class='notice'>You press firmly on the cover, dislodging it.</span>"
+					if( d_state == 2 && user.loc == T && user.get_active_hand() == W )
+						src.d_state = 3
+						src.icon_state = "r_wall-3"
+						user << "<span class='notice'>You press firmly on the cover, dislodging it.</span>"
 				return
 
 		if(3)
@@ -187,13 +187,13 @@
 				user << "<span class='notice'>You struggle to pry off the cover.</span>"
 				playsound(src, 'sound/items/Crowbar.ogg', 100, 1)
 
-				sleep(100)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,100,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( d_state == 3 && user.loc == T && user.get_active_hand() == W )
-					src.d_state = 4
-					src.icon_state = "r_wall-4"
-					user << "<span class='notice'>You pry off the cover.</span>"
+					if( d_state == 3 && user.loc == T && user.get_active_hand() == W )
+						src.d_state = 4
+						src.icon_state = "r_wall-4"
+						user << "<span class='notice'>You pry off the cover.</span>"
 				return
 
 		if(4)
@@ -202,13 +202,13 @@
 				user << "<span class='notice'>You start loosening the anchoring bolts which secure the support rods to their frame.</span>"
 				playsound(src, 'sound/items/Ratchet.ogg', 100, 1)
 
-				sleep(40)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,40,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( d_state == 4 && user.loc == T && user.get_active_hand() == W )
-					src.d_state = 5
-					src.icon_state = "r_wall-5"
-					user << "<span class='notice'>You remove the bolts anchoring the support rods.</span>"
+					if( d_state == 4 && user.loc == T && user.get_active_hand() == W )
+						src.d_state = 5
+						src.icon_state = "r_wall-5"
+						user << "<span class='notice'>You remove the bolts anchoring the support rods.</span>"
 				return
 
 		if(5)
@@ -219,14 +219,14 @@
 					user << "<span class='notice'>You begin slicing through the support rods.</span>"
 					playsound(src, 'sound/items/Welder.ogg', 100, 1)
 
-					sleep(100)
-					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !WT || !WT.isOn() || !T )	return
+					if(do_after(user,100,target = src))
+						if( !istype(src, /turf/simulated/wall/r_wall) || !user || !WT || !WT.isOn() || !T )	return
 
-					if( d_state == 5 && user.loc == T && user.get_active_hand() == WT )
-						src.d_state = 6
-						src.icon_state = "r_wall-6"
-						new /obj/item/stack/rods( src )
-						user << "<span class='notice'>The support rods drop out as you cut them loose from the frame.</span>"
+						if( d_state == 5 && user.loc == T && user.get_active_hand() == WT )
+							src.d_state = 6
+							src.icon_state = "r_wall-6"
+							new /obj/item/stack/rods( src )
+							user << "<span class='notice'>The support rods drop out as you cut them loose from the frame.</span>"
 				else
 					user << "<span class='notice'>You need more welding fuel to complete this task.</span>"
 				return
@@ -236,14 +236,14 @@
 				user << "<span class='notice'>You begin slicing through the support rods.</span>"
 				playsound(src, 'sound/items/Welder.ogg', 100, 1)
 
-				sleep(70)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,70,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( d_state == 5 && user.loc == T && user.get_active_hand() == W )
-					src.d_state = 6
-					src.icon_state = "r_wall-6"
-					new /obj/item/stack/rods( src )
-					user << "<span class='notice'>The support rods drop out as you cut them loose from the frame.</span>"
+					if( d_state == 5 && user.loc == T && user.get_active_hand() == W )
+						src.d_state = 6
+						src.icon_state = "r_wall-6"
+						new /obj/item/stack/rods( src )
+						user << "<span class='notice'>The support rods drop out as you cut them loose from the frame.</span>"
 				return
 
 		if(6)
@@ -252,12 +252,12 @@
 				user << "<span class='notice'>You struggle to pry off the outer sheath.</span>"
 				playsound(src, 'sound/items/Crowbar.ogg', 100, 1)
 
-				sleep(100)
-				if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+				if(do_after(user,100,target = src))
+					if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-				if( user.loc == T && user.get_active_hand() == W )
-					user << "<span class='notice'>You pry off the outer sheath.</span>"
-					dismantle_wall()
+					if( user.loc == T && user.get_active_hand() == W )
+						user << "<span class='notice'>You pry off the outer sheath.</span>"
+						dismantle_wall()
 				return
 
 //vv OK, we weren't performing a valid deconstruction step or igniting thermite,let's check the other possibilities vv
@@ -267,12 +267,12 @@
 
 		user << "<span class='notice'>You begin to drill though the wall.</span>"
 
-		sleep(200)
-		if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
+		if(do_after(user,200,target = src))
+			if( !istype(src, /turf/simulated/wall/r_wall) || !user || !W || !T )	return
 
-		if( user.loc == T && user.get_active_hand() == W )
-			user << "<span class='notice'>Your drill tears though the last of the reinforced plating.</span>"
-			dismantle_wall()
+			if( user.loc == T && user.get_active_hand() == W )
+				user << "<span class='notice'>Your drill tears though the last of the reinforced plating.</span>"
+				dismantle_wall()
 
 	//REPAIRING
 	else if( istype(W, /obj/item/stack/sheet/metal) && d_state )
@@ -280,18 +280,18 @@
 
 		user << "<span class='notice'>You begin patching-up the wall with \a [MS].</span>"
 
-		sleep( max(20*d_state,100) )	//time taken to repair is proportional to the damage! (max 10 seconds)
-		if( !istype(src, /turf/simulated/wall/r_wall) || !user || !MS || !T )	return
+		if(do_after(user,(max(20*d_state,100)),target = src))	//time taken to repair is proportional to the damage! (max 10 seconds)
+			if( !istype(src, /turf/simulated/wall/r_wall) || !user || !MS || !T )	return
 
-		if( user.loc == T && user.get_active_hand() == MS && d_state )
-			src.d_state = 0
-			src.icon_state = "r_wall"
-			relativewall_neighbours()	//call smoothwall stuff
-			user << "<span class='notice'>You repair the last of the damage.</span>"
-			if (MS.amount > 1)
-				MS.amount--
-			else
-				qdel(MS)
+			if( user.loc == T && user.get_active_hand() == MS && d_state )
+				src.d_state = 0
+				src.icon_state = "r_wall"
+				relativewall_neighbours()	//call smoothwall stuff
+				user << "<span class='notice'>You repair the last of the damage.</span>"
+				if (MS.amount > 1)
+					MS.amount--
+				else
+					qdel(MS)
 
 	//APC
 	else if( istype(W,/obj/item/apc_frame) )
